backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
joblib==1.3.2